Add detailed spritesheet tests
These tests will loop through all icons defined in a spritesheet and
make sure they actually exist in the resulting PNG file. Sharp will
throw an error if the sprite is out of bounds or missing.

Maybe this test should be improved to make sure the contents of the cut
sprite actually match the source files, but I decided that was out of
scope for now (since it would be rendering SVG, there could be minor
differences so a byte-for-byte check wouldn't be feasible)

This commit also adds Maki icons as a test set.